🔧 refactor: Organize repository-related code and extract types
- Moved `getRepoInfo` and `RepoInfo` interface to a new `lib/methods.ts` and `types/objects.ts` files for better structure and reusability.
- Removed redundant code from `page.tsx` and `pr-list.tsx` by importing shared types and functions.
- Streamlined the `RepoHome` component to utilize the new imports, enhancing maintainability.
